Noncommutative polynomial optimization

TSSOS supports noncommutative polynomial (eigenvalue or trace) optimization involving noncommuting variables. See also NCTSSOS.

TSSOS.NCTSSOS.nctssos_first — Function

opt,data = nctssos_first(f::Polynomial{false, T} where T<:Number, x::Vector{PolyVar{false}};
    newton=true, reducebasis=true, TS="block", obj="eigen", merge=false, md=3, solve=true, QUIET=false)

Compute the first step of the NCTSSOS hierarchy for unconstrained noncommutative polynomial optimization. If newton=true, then compute a monomial basis by the Newton chip method. If reducebasis=true, then remove monomials from the monomial basis by diagonal inconsistency. If TS="block", use maximal chordal extensions; if TS="MD", use approximately smallest chordal extensions. If obj="eigen", minimize the eigenvalue; if obj="trace", then minimize the trace. If merge=true, perform the PSD block merging. Return the optimum and other auxiliary data.

Arguments

f: the objective function for unconstrained noncommutative polynomial optimization.

x: the set of noncommuting variables.

md: the tunable parameter for merging blocks.

TSSOS.NCTSSOS.nctssos_higher! — Function

opt,data = nctssos_higher!(data, TS="block", merge=false, md=3, QUIET=false)

Compute higher steps of the NCTSSOS hierarchy. Return the optimum and other auxiliary data.

TSSOS.NCTSSOS.cs_nctssos_first — Function

opt,data = cs_nctssos_first(pop, x, d; numeq=0, CS="MF", TS="block", merge=false, md=3,
QUIET=false, obj="eigen", solve=true)

Compute the first step of the CS-NCTSSOS hierarchy for constrained noncommutative polynomial optimization with relaxation order d. Return the optimum and other auxiliary data.

Arguments

pop: the vector of the objective function, inequality constraints, and equality constraints.

x: the set of noncommuting variables.

d: the relaxation order of the moment-SOHS hierarchy.

numeq: the number of equality constraints.

opt,data = cs_nctssos_first(supp::Vector{Vector{Vector{UInt16}}}, coe, n::Int, d::Int; numeq=0,
CS="MF", TS="block", merge=false, md=3, QUIET=false, obj="eigen", solve=true)

Compute the first step of the CS-NCTSSOS hierarchy for constrained noncommutative polynomial optimization with relaxation order d. Here the polynomial optimization problem is defined by supp and coe, corresponding to the supports and coeffients of pop respectively. Return the optimum and other auxiliary data.

Arguments

supp: the supports of the polynomial optimization problem.

coe: the coeffients of the polynomial optimization problem.

d: the relaxation order of the moment-SOHS hierarchy.

numeq: the number of equality constraints.

TSSOS.NCTSSOS.cs_nctssos_higher! — Function

opt,data = cs_nctssos_higher!(data; TS="block", QUIET=false, merge=false, md=3, solve=true)

Compute higher steps of the CS-NCTSSOS hierarchy. Return the optimum and other auxiliary data.