Polynomial optimization

TSSOS.tssos_first — Function

opt,sol,data = tssos_first(f, x; nb=0, newton=true, reducebasis=false, TS="block", merge=false,
md=3, feasible=false, solver="Mosek", QUIET=false, solve=true, MomentOne=false, Gram=false, solution=false, tol=1e-4)

Compute the first TS step of the TSSOS hierarchy for unconstrained polynomial optimization. If newton=true, then compute a monomial basis by the Newton polytope method. If reducebasis=true, then remove monomials from the monomial basis by diagonal inconsistency. If TS="block", use maximal chordal extensions; if TS="MD", use approximately smallest chordal extensions. If merge=true, perform the PSD block merging. If feasible=true, then solve the feasibility problem. If solve=false, then do not solve the SDP. If Gram=true, then output the Gram matrix. If MomentOne=true, add an extra first order moment matrix to the moment relaxation.

Input arguments

f: objective

x: POP variables

nb: number of binary variables in x

md: tunable parameter for merging blocks

QUIET: run in the quiet mode or not (true, false)

tol: relative tolerance to certify global optimality

Output arguments

opt: optimum

sol: (near) optimal solution (if solution=true)

data: other auxiliary data

opt,sol,data = tssos_first(pop, x, d; nb=0, numeq=0, quotient=true, basis=[],
reducebasis=false, TS="block", merge=false, md=3, solver="Mosek", QUIET=false, solve=true,
MomentOne=false, Gram=false, solution=false, tol=1e-4)

Compute the first TS step of the TSSOS hierarchy for constrained polynomial optimization. If quotient=true, then exploit the quotient ring structure defined by the equality constraints. If merge=true, perform the PSD block merging. If solve=false, then do not solve the SDP. If Gram=true, then output the Gram matrix. If MomentOne=true, add an extra first order moment matrix to the moment relaxation.

Input arguments

pop: vector of the objective, inequality constraints, and equality constraints

x: POP variables

d: relaxation order

nb: number of binary variables in x

numeq: number of equality constraints

TS: type of term sparsity ("block", "MD", "MF", false)

md: tunable parameter for merging blocks

QUIET: run in the quiet mode or not (true, false)

tol: relative tolerance to certify global optimality

Output arguments

opt: optimum

sol: (near) optimal solution (if solution=true)

data: other auxiliary data

TSSOS.tssos_higher! — Function

opt,sol,data = tssos_higher!(data; TS="block", merge=false, md=3, QUIET=false, solve=true,
MomentOne=false, solution=false, tol=1e-4)

Compute higher TS steps of the TSSOS hierarchy.

TSSOS.cs_tssos_first — Function

opt,sol,data = cs_tssos_first(pop, x, d; nb=0, numeq=0, CS="MF", cliques=[], TS="block", merge=false, md=3, solver="Mosek", QUIET=false, solve=true, solution=false,
Gram=false, MomentOne=true, Mommat=false, tol=1e-4)

Compute the first TS step of the CS-TSSOS hierarchy for constrained polynomial optimization. If merge=true, perform the PSD block merging. If solve=false, then do not solve the SDP. If Gram=true, then output the Gram matrix. If Mommat=true, then output the moment matrix. If MomentOne=true, add an extra first order moment matrix to the moment relaxation.

Input arguments

pop: vector of the objective, inequality constraints, and equality constraints

x: POP variables

d: relaxation order

nb: number of binary variables in x

numeq: number of equality constraints

CS: method of chordal extension for correlative sparsity ("MF", "MD", false)

cliques: the set of cliques used in correlative sparsity

TS: type of term sparsity ("block", "MD", "MF", false)

md: tunable parameter for merging blocks

QUIET: run in the quiet mode or not (true, false)

tol: relative tolerance to certify global optimality

Output arguments

opt: optimum

sol: (near) optimal solution (if solution=true)

data: other auxiliary data

opt,sol,data = cs_tssos_first(supp::Vector{Vector{Vector{UInt16}}}, coe, n, d; nb=0, numeq=0, CS="MF", cliques=[], TS="block", 
merge=false, md=3, QUIET=false, solver="Mosek", solve=true, solution=false, Gram=false, MomentOne=true, Mommat=false, tol=1e-4)

Compute the first TS step of the CS-TSSOS hierarchy for constrained polynomial optimization. Here the polynomial optimization problem is defined by supp and coe, corresponding to the supports and coeffients of pop respectively.

opt,sol,data = cs_tssos_first(pop, z, n, d; nb=0, numeq=0, CS="MF", cliques=[], TS="block", ipart=true, merge=false, md=3, 
solver="Mosek", QUIET=false, solve=true, Gram=false, Mommat=false, MomentOne=true)

Compute the first TS step of the CS-TSSOS hierarchy for constrained complex polynomial optimization. If merge=true, perform the PSD block merging. If ipart=false, then use the real moment-HSOS hierarchy. If solve=false, then do not solve the SDP. If Gram=true, then output the Gram matrix. If Mommat=true, then output the moment matrix. If MomentOne=true, add an extra first order moment matrix to the moment relaxation.

Input arguments

pop: vector of the objective, inequality constraints, and equality constraints

z: CPOP variables and their conjugate

n: number of CPOP variables

d: relaxation order

nb: number of unit-norm variables in x

numeq: number of equality constraints

CS: method of chordal extension for correlative sparsity ("MF", "MD", false)

cliques: the set of cliques used in correlative sparsity

TS: type of term sparsity ("block", "MD", "MF", false)

md: tunable parameter for merging blocks

QUIET: run in the quiet mode or not (true, false)

Output arguments

opt: optimum

data: other auxiliary data

opt,sol,data = cs_tssos_first(supp::Vector{Vector{Vector{Vector{UInt16}}}}, coe::Vector{Vector{ComplexF64}},
n, d; nb=0, numeq=0, CS="MF", cliques=[], TS="block", ipart=true, merge=false, md=3, solver="Mosek",
QUIET=false, solve=true, Gram=false, Mommat=false, MomentOne=true)

Compute the first TS step of the CS-TSSOS hierarchy for constrained complex polynomial optimization. Here the complex polynomial optimization problem is defined by supp and coe, corresponding to the supports and coeffients of pop respectively.

TSSOS.cs_tssos_higher! — Function

opt,sol,data = cs_tssos_higher!(data; TS="block", merge=false, md=3, QUIET=false, solve=true,
solution=false, Gram=false, Mommat=false, MomentOne=false)

Compute higher TS steps of the CS-TSSOS hierarchy.

TSSOS.refine_sol — Function

ref_sol,flag = refine_sol(opt, sol, data, QUIET=false, tol=1e-4)

Refine the obtained solution by a local solver. Return the refined solution, and flag=0 if global optimality is certified, flag=1 otherwise.