Polynomial optimization

TSSOS.tssos_first — Function

opt,sol,data = tssos_first(f, x; nb=0, newton=true, reducebasis=false, TS="block", merge=false,
md=3, solver="Mosek", QUIET=false, solve=true, MomentOne=false, Gram=false, solution=false, tol=1e-4)

Compute the first step of the TSSOS hierarchy for unconstrained polynomial optimization. If newton=true, then compute a monomial basis by the Newton polytope method. If reducebasis=true, then remove monomials from the monomial basis by diagonal inconsistency. If TS="block", use maximal chordal extensions; if TS="MD", use approximately smallest chordal extensions. If merge=true, perform the PSD block merging. If MomentOne=true, add an extra first order moment matrix to the moment relaxation. Return the optimum, the (near) optimal solution (if solution=true) and other auxiliary data.

Arguments

f: the objective function for unconstrained polynomial optimization.

x: the set of variables.

nb: the number of binary variables in x.

md: the tunable parameter for merging blocks.

tol: the relative tolerance to certify global optimality.

opt,sol,data = tssos_first(pop, x, d; nb=0, numeq=0, quotient=true, basis=[],
reducebasis=false, TS="block", merge=false, md=3, solver="Mosek", QUIET=false, solve=true,
MomentOne=false, Gram=false, solution=false, tol=1e-4)

Compute the first step of the TSSOS hierarchy for constrained polynomial optimization with relaxation order d. If quotient=true, then exploit the quotient ring structure defined by the equality constraints. Return the optimum, the (near) optimal solution (if solution=true) and other auxiliary data.

Arguments

pop: the vector of the objective function, inequality constraints, and equality constraints.

x: the set of variables.

d: the relaxation order of the moment-SOS hierarchy.

nb: the number of binary variables in x.

md: the tunable parameter for merging blocks.

numeq: the number of equality constraints.

TSSOS.tssos_higher! — Function

opt,sol,data = tssos_higher!(data; TS="block", merge=false, md=3, QUIET=false, solve=true,
MomentOne=false, solution=false, tol=1e-4)

Compute higher steps of the TSSOS hierarchy. Return the optimum, the (near) optimal solution (if solution=true) and other auxiliary data.

TSSOS.cs_tssos_first — Function

opt,sol,data = cs_tssos_first(pop, x, d; nb=0, numeq=0, foc=100, CS="MF", minimize=false,
assign="first", TS="block", merge=false, md=3, solver="Mosek", QUIET=false, solve=true, solution=false,
MomentOne=true, tol=1e-4)

Compute the first step of the CS-TSSOS hierarchy for constrained polynomial optimization with relaxation order d. Return the optimum, the (near) optimal solution (if solution=true) and other auxiliary data.

Arguments

pop: the vector of the objective function, inequality constraints, and equality constraints.

x: the set of variables.

d: the relaxation order of the moment-SOS hierarchy.

nb: the number of binary variables in x.

md: the tunable parameter for merging blocks.

numeq: the number of equality constraints.

opt,sol,data = cs_tssos_first(supp::Vector{Vector{Vector{UInt16}}}, coe, n, d; numeq=0,
nb=0, foc=100, CS="MF", minimize=false, assign="first", TS="block", merge=false, md=3,
QUIET=false, solver="Mosek", solve=true, solution=false, MomentOne=true, tol=1e-4)

Compute the first step of the CS-TSSOS hierarchy for constrained polynomial optimization with relaxation order d. Here the polynomial optimization problem is defined by supp and coe, corresponding to the supports and coeffients of pop respectively. Return the optimum, the (near) optimal solution (if solution=true) and other auxiliary data.

Arguments

supp: the supports of the polynomial optimization problem.

coe: the coeffients of the polynomial optimization problem.

d: the relaxation order of the moment-SOS hierarchy.

numeq: the number of equality constraints.

opt,sol,data = cs_tssos_first(supp::Vector{Vector{Vector{Vector{UInt16}}}}, coe::Vector{Vector{ComplexF64}},
n, d; numeq=0, foc=100, CS="MF", minimize=false, assign="first", TS="block", merge=false, md=3, solver="Mosek",
QUIET=false, solve=true, MomentOne=true)

Compute the first step of the CS-TSSOS hierarchy for constrained complex polynomial optimization with relaxation order d. Here the complex polynomial optimization problem is defined by supp and coe, corresponding to the supports and coeffients of pop respectively.

Arguments

supp: the supports of the complex polynomial optimization problem.

coe: the coeffients of the complex polynomial optimization problem.

d: the relaxation order of the moment-SOHS hierarchy.

numeq: the number of equality constraints.

TSSOS.cs_tssos_higher! — Function

opt,sol,data = cs_tssos_higher!(data; TS="block", merge=false, md=3, QUIET=false, solve=true,
solution=false, MomentOne=false)

Compute higher steps of the CS-TSSOS hierarchy. Return the optimum, the (near) optimal solution (if solution=true) and other auxiliary data.

TSSOS.refine_sol — Function

ref_sol,upper_bound,rel_gap = refine_sol(opt, sol, data, QUIET=false, tol=1e-4)

Refine the obtained solution by a local solver. Return the refined solution, the upper bound given by the local solver and the relative optimality gap.